chore: guard markup traversal imports
